Moving the meta model related file to different directory
