+++ turbo_ml/forecast/__init__.py
@@ -1,4 +1,4 @@
 from .model_forecast import Forecast, ExhaustiveSearch
 from .hyperparameter_tuning import HyperTuner
 from .dataset_characteristics import StatisticalParametersExtractor
-from .meta_model_search import MetaModelGuesser
+from ..meta_model.meta_model_search import MetaModelGuesser
