+++ turbo_ml/utils/options.py
@@ -3,11 +3,12 @@ from typing import Literal
 
 
 class options:
-    print_logs: bool = True  # TODO: change default to False before release
+    print_logs: bool = True
     # TODO: change default to auto-adjustem based on terminal size before release
     text_size: int = 64
     validation: bool = True  # TODO: change default to False before release
-    device: Literal['cpu', 'cuda', 'mps'] = 'cuda'
+    # TODO: change default to automatic detection
+    device: Literal['cpu', 'cuda', 'mps'] = 'cpu'
     threads: int = -1
     dev_mode = True
     dev_mode_logging = logging.INFO
